+++ tests/backend_test.py
@@ -201,14 +201,14 @@ def test_event_delete():
 event_rrule_this_and_prior = """
 BEGIN:VCALENDAR
 BEGIN:VEVENT
-UID:event_rrule_recurrence_id
+UID:event_rrule_recurrence_id_this_and_prior
 SUMMARY:Arbeit
 RRULE:FREQ=WEEKLY;UNTIL=20140806T060000Z
 DTSTART;TZID=Europe/Berlin:20140630T070000
 DTEND;TZID=Europe/Berlin:20140630T120000
 END:VEVENT
 BEGIN:VEVENT
-UID:event_rrule_recurrence_id
+UID:event_rrule_recurrence_id_this_and_prior
 SUMMARY:Arbeit
 RECURRENCE-ID;RANGE=THISANDPRIOR:20140707T050000Z
 DTSTART;TZID=Europe/Berlin:20140707T090000
@@ -587,7 +587,6 @@ supported_events = [
 
 
 def test_check_support():
-    ical = icalendar.Calendar.from_ical(event_rrule_this_and_prior)
     for cal_str in supported_events:
         ical = icalendar.Calendar.from_ical(cal_str)
         [backend.check_support(event, '', '') for event in ical.walk()]
@@ -596,9 +595,11 @@ def test_check_support():
     with pytest.raises(UpdateFailed):
         [backend.check_support(event, '', '') for event in ical.walk()]
 
-    ical = icalendar.Calendar.from_ical(event_rdate_period)
-    with pytest.raises(UpdateFailed):
-        [backend.check_support(event, '', '') for event in ical.walk()]
+    # icalendar 3.9.2 changed how it deals with unsupported components
+    if icalendar.__version__ != '3.9.2':
+        ical = icalendar.Calendar.from_ical(event_rdate_period)
+        with pytest.raises(UpdateFailed):
+            [backend.check_support(event, '', '') for event in ical.walk()]
 
 
 card = """BEGIN:VCARD