+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ not released yet
 
 * BREAKING CHANGE: python 2 is no longer supported (Hugo Osvaldo Barrera)
 * updated dependency: vdirsyncer >= 0.5.2
+* make tests work with icalendar 3.9.2 (no functional changes) (Christian Geier)
 
 * support for showing the birthday of contacts with no FN property (Hugo
   Osvaldo Barrera)